photo_anotation_util/requirements.txt
opencv-python>=4.8.0
Pillow>=10.0.0
rawpy>=0.17.0